refactor: extract duplicated years fetching logic into reusable utility
- Create fetchAvailableFinanceYears utility function in financeCalculations.ts
- Remove duplicated pagination logic from API route and page component
- Maintain same functionality with RPC fallback and pagination support
- Improve maintainability by centralizing business logic
- Reduce risk of inconsistencies between implementations
