+++ app/(dashboard)/finanzen/page.tsx
@@ -112,76 +112,8 @@ async function getSummaryDataFallback(year: number) {
 
 async function getAvailableYears() {
   const supabase = await createClient();
-  
-  try {
-    // Try to use the optimized database function
-    const { data, error } = await supabase.rpc('get_available_finance_years');
-    
-    if (!error && data) {
-      return data.map((item: any) => item.year).sort((a: number, b: number) => b - a);
-    }
-  } catch (error) {
-    console.log('RPC function not available for years, using fallback');
-  }
-
-  // Fallback to regular query with pagination
-  const currentYear = new Date().getFullYear();
-  const years = new Set<number>();
-  
-  // Add current year by default
-  years.add(currentYear);
-  
-  try {
-    const pageSize = 1000;
-    let page = 0;
-    let hasMore = true;
-
-    while (hasMore) {
-      const { data, error } = await supabase
-        .from('Finanzen')
-        .select('datum')
-        .not('datum', 'is', null)
-        .range(page * pageSize, (page + 1) * pageSize - 1);
-
-      if (error) {
-        console.error('Error fetching available years:', error);
-        break;
-      }
-
-      if (!data || data.length === 0) {
-        hasMore = false;
-        break;
-      }
-
-      // Process dates to extract years
-      data.forEach(item => {
-        if (!item.datum) return;
-        
-        try {
-          const date = new Date(item.datum);
-          if (!isNaN(date.getTime())) {
-            const year = date.getFullYear();
-            if (year <= currentYear + 1) {
-              years.add(year);
-            }
-          }
-        } catch (e) {
-          console.warn('Invalid date format:', item.datum);
-        }
-      });
-
-      // If we got fewer records than the page size, we've reached the end
-      if (data.length < pageSize) {
-        hasMore = false;
-      } else {
-        page++;
-      }
-    }
-  } catch (error) {
-    console.error('Error in pagination fallback for available years:', error);
-  }
-
-  return Array.from(years).sort((a, b) => b - a);
+  const { fetchAvailableFinanceYears } = await import("@/utils/financeCalculations");
+  return await fetchAvailableFinanceYears(supabase);
 }
 
 export default async function FinanzenPage() {
